backend/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
flask-migrate
marshmallow-sqlalchemy
psycopg2